feat(config): add support for specifying port
Often times you might need to configure a custom host or IP (e.g. for a
proxy) and you might then need to specify a port for that connection to
function. This adds support for that. 

**Note:** This removes the validations of various hosts (e.g. you can
specify any host, including an ip address), but it still supports using
`--host foo` and having that expand to `foo-be.glean.com`.

- Add port field to ConfigOptions struct
- Update CLI flags and help text to include port option
- Modify HTTP client to use port in base URL construction
- Add support for foo.bar.com host format
- Add tests for port configuration and URL building